Update
> Se creo la clase Skeletal Mesh, falta implementacion adecuada.
> Se agregaron los pragmas necesarios para quitar errores de DX.
> Cambios de nombres en giPlatformMath.h -> todas las funciones sqr -> sqrt.
> Se agregaron las rotaciones de matrices en los ejes x y z.
> Se cambio el tipo de puntero del Timer en giTime.h.
> Se corrigieron errores en la matematica de los vectores4i.
> Se genero la clase Material.
> Se eliminaron las animaciones de las clases Model & Mesh.
> Se termino de implementar el Scene graph y sus nodos.
> Se realizaron algunos cambios en algunos tipos de datos de algunos objetos.
> Se elimino la carga del cubo a mano del main.

+++ giCore/include/giBaseApp.h
@@ -146,10 +146,10 @@ class GI_CORE_EXPORT BaseApp
   //static LRESULT CALLBACK handleWindowEvent(HWND inHw, UINT msg, WPARAM wParam, LPARAM lParam);
 
   ///The Width of the window with a default size
-  float m_width = 640.f;
+  uint32 m_width = 640.f;
   
   ///The Height of the window with a default size
-  float m_height = 480.f;
+  uint32 m_height = 480.f;
   
   ///The reference of the window
   WindowBase m_window;
